Fix JRuby compatibility
See the comments on https://github.com/avdi/naught/pull/27

- Ran into prolems identifying the line and file of instantiation
for traceable null objects because of slight differences in
MRI's and JRuby's backtraces.
    - Create a RubyEngine module to handle slight differences like
the above between different ruby engines.
- JRuby's backtrace doesn't give a clear indication of when methods
are called from inside blocks.
    - It's not vital that this information be tracked by
pebble objects, so we'll drop the specs that try to verify that.
- JRuby's backtrace also prints out a different trace for methods:
"from calling_method" instead of "in calling_method".
    - Drop the preposition and just verify against the method name