Move dependencies from gemspec to Gemfile (expect for Bundler)
Bundler's Gemfile gives you the ability to differentiate between
development dependencies and test dependencies. Non-test development
dependencies (e.g. guard) do not need to be installed in the continuous
integration environment, making the build somewhat faster.

This leaves bundler as the only dependency specified in the gemspec,
since it is needed to bootstrap dependencies specified in the Gemfile.